Moved org.javacs.kt.classpath and .util to 'shared' module and extracted PsiUtils from Utils.kt
