+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'KotlinLanguageServer'
 
-include 'server'
+include 'shared', 'server'
