+++ shared/src/main/kotlin/org/javacs/kt/util/AsyncExecutor.kt
