+++ shared/src/main/kotlin/org/javacs/kt/util/DelegatePrintStream.kt
