+++ shared/src/main/kotlin/org/javacs/kt/util/ExitingInputStream.kt
