+++ shared/src/main/kotlin/org/javacs/kt/util/KotlinLSException.kt
